Front: CSS on post button ok

+++ front/src/components/ButtonDelete.vue
@@ -16,19 +16,29 @@ export default {
 	},
 	methods: {
 		async deleteCard(table) {
-			await axios.delete(`${table}/${this.$props.id}`)
-			table == "comments" ? location.reload() : this.$router.push("/")
+			if (confirm("Are you sure you want to delete this post?")) {
+				await axios.delete(`${table}/${this.$props.id}`)
+				table == "comments" ? location.reload() : this.$router.push("/")
+				location.reload()
+				alert("Your post has been deleted")
+			}
 		}
 	}
 }
 </script>
 <template>
-	<button
-		@click="deleteCard(this.table)"
-		type="button"
-		class="postCard__button--delete"
-	>
-		Delete
-	</button>
+	<div>
+		<img
+			class="trashcan-button"
+			src="../assets/trashcan.svg"
+			alt="Delete button"
+			@click="deleteCard(this.table)"
+		/><span class="button-label">Delete</span>
+	</div>
 </template>
-<style></style>
+<style scoped lang="scss">
+.trashcan-button {
+	height: 20px;
+	cursor: pointer;
+}
+</style>
